refactor: re-structure the code
